requirements.txt
litellm>=1.22.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
roo-cline>=1.0.0  # Add roo-cline for testing
pytest-cov>=4.1.0  # For coverage reporting
fastapi>=0.109.0
uvicorn>=0.25.0
jinja2>=3.1.2
python-multipart>=0.0.6
